requirements.txt
httpx>=0.27
pandas>=2.0
pydantic>=2.0
python-dotenv>=1.0
